fix: replace lock dependency with uv lower installation
As having an optional dependency to host lower constraints lockfile
caused problems with renovate and dependabot, we avoid it by relying
on uv lower resolver for testing our lower limits.

This also required adding few extra dependency conditions for indirect
dependencies that would create conflicts otherwise.

This approach seems to be better because it already spotted few bugs,
like setuptools not requiring a minimal version and same with uv.